I added a section on moving and resizing and another on scrolling.
I am not particularly proud of the structure of the manual at this
point.  The two new sections are in the chapter "CLIM drawing
functions" which doesn't seem right.  However, I thought it would be
better to write what I did while I remember the issues and restructure
later.

If you have some time, please read the new sections and give me
feedback.  They contain information that is not explicitly in the
specification, so part of the information is my own interpretation of
the specification.  Getting this information right is essential to
getting the implementation right.

Oh, and I also fixed the terminology so that "mirrored sheet" means a
sheet that actually _has_ a mirror, as opposed to an instance of the
class mirrored-sheet-mixin.