Reorganized the chapters somewhat.
Worked on sections explaining how to compute the native transformation
and the native region.  I know think we need a generic function

   mirror-transformation port mirror

that works like sheet-transformation, but that works on mirrors, i.e.,
it returns a transformation that translates coordinates in the mirror
to coordinates in the mirrors parent.  Methods will typically specialize
on the port.  For the CLX backend, the transformation returned will
always be a translation.