Rewrote the copy-{from,to}-pixmap and copy-area functions according to the new spec.
The main difference is that the three functions now rely on
medium-copy-area, which is specialized on both a source an a destination
`drawable' and is in charge of the low-level stuff, such as
xlib:copy-area for clx-medium.  This leads to clearer and shorter code.

The copy to/from a stream is not implemended yet.