Fixed bug reported by Paul Werkowski; coordinate sequences can now be lists or vectors.  Introduced a do-sequence macro that has similar semantics to dotimes and dolist as well as features of map-sequence.
Implemented frame-state (except for :shrunk).  Changed timing of when
frames are generated, adopted and enabled to follow the spec.  Fixed
clim-fig to adapt to this new world order.  Disable the frame when the
top level function exits.

Remove erroring methods for copy-to-pixmap (stream ...) and friends.

Wrote a with-keywords-removed macro that attempts to be efficient
about about removing arguments from argument lists.