Got OPEN-WINDOW-STREAM working.
There are a few things about window-streams that are still not perfect:

Output in the window stream does not seem to influence the scroll bars.
The existing code initialized the interior pane to an arbitrary size and
I've left it that way. This needs to be fixed.

Apparently the common mode of using window-streams was to have them share
the input buffer of their parent window. In this case, what to do when the
parent closes?

For window-streams with their own input buffer, I start a new thread and
run a minimal toplevel to handle events. If the window is instead sharing
the input buffer, shutdown of the main *application-pane*'s toplevel will
leave the window(s) in limbo with no one handling their events. There are
maybe three approaches to handling this:

1) Assume the parent window has responsibility for closing any window streams
   it opens. (bleh)
2) Record them somewhere and close them.
3) Make them work automagically.

To implement #3, we'll have to keep track of these window streams somewhere,
and arrange to have an event handling thread started in the case of the main
toplevel shutting down. A good argument against NOT using the :input-buffer
argument to OPEN-WINDOW-STREAM. Something to think about.