Implemented command-or-form presentation type. Fixed stupid-subtypep so OR presentation types can be used in with-input-context. Changed the subform reader stuff to use a presentation type option instead of the presentation type subform. Got rid of subform altogether. Changed the Lisp presentation types to be a subtype of expression, not form; added a presentation translator from expression to form that does necessary quoting.
Fixed passing of arguments to the frame-top-level function. Moved the
menu-item stuff in read-frame-command into an :around method so that
read-frame-command can be usefully overridden by users.

Cleaned up the listener demo: got rid of the top level loop parts that
are now redundant. Added a frame slot and an argument to
run-listener/run-listener-process to control whether or not to use CLIM
command-or-form processing (the default is the existing method.)