Implement :multiple-window option to tracking-pointer. Added a new backend protocol to support this, through functions PORT-GRAB-POINTER and PORT-UNGRAB-POINTER.
Using this, ensure that MENU-CHOOSE grabs the pointer and clicking anywhere
outside the menu will dismiss it (which previously only worked somewhat,
due to the lack of pointer grabbing).

Note the slight limitation that the CLX backend does not always accurately
report the window which the pointer is over during a grab.

Imported some symbols into CLIM-CLX package.

Changed the Help command to present available commands using the item
formatter.