Fix hackish redisplay of wholine pane in response to paint events (now redisplay in response to allocate-space instead).
Eliminated remnants of listener top-level function.

Added several keywords to run-listener. :new-process controls whether the
listener will run in the current or a new process. :eval takes a form or
function and executes it within the listener at startup. :process-name
sets the name of the new process, if one is created.

Eliminated run-listener-process function since this can now be done with
(run-listener :new-process t).