In find-applicable-translators, call remove-duplicates on the resulting list so that commands reached multiple times via inheritance only appear once (fixing duplicate command menu items in the listener).
Remove comment in menu.lisp that only :command and :menu items are
supported, which has changed.

In the main CLX event handler for configure-notify events, use
xlib:translate-coordinates to compute the position of the top-level sheet
relative to the screen root window. This fixes a problem where resizing a
frame caused the sheet-delta-transformation to be incorrect, causing
menus from the menubar to appear in incorrect places.