Implement :SINGLE-BOX properly.:SINGLE-BOX NIL is the default, but McCLIM has ignored it. This changes (for the better) the behavior of applications.
Implement PRESENTATION-REFINED-POSITION-TEST.

If the view argument to ACCEPT is a list, apply MAKE-INSTANCE to the
list to obtain the view.

Fix a bug in dialogs that prevented moving on to the next text field
after hitting return.

Define a new view type, TEXT-FIELD-VIEW, that is used in
dialogs. This view has a WIDTH parameter.