The inspector now uses monospaced fonts for certain things, which looks nicer.
Formatting of package lists is cleaned up, with the addition of a new
function to display a list vertically without parentheses.

Functions can now be disassembled, and the disassembly can be
displayed in implementation-specific ways. Currently the only
implementation that has a specific format is SBCL, and it might be
broken on more recent versions if they've changed the disassembly
format significantly since 0.8.16 without telling me.