Commit working version of text-bounding-rectangle* stuff, as trailed on mcclim-devel 2005-08-12.  Basically, this is needed because the drawn area for left-to-right text need not lie between (x,y-ascent) and (x+width,y+descent).  (No doubt other text drawing directions suffer from the same problem, but they're not yet implemented in McCLIM).
New per-medium function TEXT-BOUNDING-RECTANGLE* which actually returns
the bounding-rectangle* of what is drawn (cf. TEXT-WIDTH, which doesn't
do anything of the sort).  Use it in DEF-GRECORDING DRAW-TEXT, and in
add-{string,character}-to-output-record, to properly adjust the output
record coordinates.  While we're at it, fix the bounding box for
:y-align :center DRAW-TEXT.

Implement this per-medium function for the CLX backend, for the
experimental freetype text handling, and for postscript (tested using
Climacs, Tabcode and clim-demo::postscript-text).  This patch was mostly
motivated by the observation that incremental redisplay in climacs
windows using the freetype backend caused graphical artifacts to appear
over time, thanks to glyphs drawing outside the output record bounding
rectangles.

Breaks:
 * CLX backend with #+unicode (clisp?)
 * Beagle backend
 * OpenGL backend
(please fix!)