This patch HANDLER-BINDs the ABORT-GESTURE condition to #'abort for the function ACCEPT in presentation-defs.lisp.
ABORT-GESTURE is the condition that is signaled when any of the
gestures in *ABORT-GESTURES* is read (in STREAM-READ-GESTURE).
Right now *ABORT-GESTURES* contains only :abort on mcclim, which
is a the keyboard gesture (#\c :control) (on Genera it contains
#\Abort, the ABORT-key).

I do not find explicitly in the clim specification that an ACCEPT
should be aborted on an ABORT-GESTURE, but it seems to be the
right thing (and I have to admit that I haven't been looking very
hard).

I did short tests with ACCEPTING-VALUES and it seems to behave
correctly with this patch, i.e. the whole dialog will be aborted.
But perhaps it would be nicer if, as long as a gadget of the
dialog is selected, only the edit of that gadget were aborted.