Section 23.7.1 "Defining Presentation Translators" of the clim spec says in the subsection on DEFINE-PRESENTATION-ACTION:
    Note that the tester for presentation actions is always assumed to be
    definitive.

This hasn't been the case until now for McCLIM.

In addition to this the lambda-list of DEFINE-DRAG-AND-DROP-TRANSLATOR
has been revised a bit. (Note that I couldn't find a sentence in the
spec that specifies the default value for the parameter MENU, but the
other three define-translator macros have all t as the default.
Nevertheless, most of the parameters are ignored right now.)