The commit on 2006-03-01 to clim-area claimed to add space for a visible cursor, but in fact added space for a cursor whether visible or not. Only add in the space for visible cursors.
This fixes most of the text glitches in the clim listener; now
where previously e.g. (* 2 3) RET would leave an extra space at the end
of the bounding rectangle, the text is tightly bound.  *package* RET
still has a problem, as the RET there seems to open a new (empty) screen
line.