Magic number elimination #2:
In sheet-mirror-region, do what the comment already claimed it does:
Ask the backend for mirror sizes if the frontend does not have a cached
value.

Previously, width and height were returned as 65536, a completely bogus
value.  This was presumably because the CLX backend started out with the
too-small 100x100 size.