Improved Drei redisplay performance by 66% in most cases.
The main difference is that syntaxes are now supposed to report which
parts of the display may need to be updated, previously their view of
the display was computed for every redisplay iteration, and any
changes drawn.

Of course, no syntaxes do that yet, so if you use Lisp block-comments
or string-quoting, you will see "delayed" redrawing of some parts of
the display. Just like Emacs!

Currently, a heuristic is used that invalidates parts of the display
corresponding to buffer regions that have actually been changed, so it
does work fine for the common cases.