Moved macro WITH-MEDIUM-OPTIONS to design.lisp.
The macro WITH-MEDIUM-OPTIONS was previously defined in graphics.lisp.
However, graphics.lisp depends on design.lisp, and WITH-MEDIUM-OPTIONS
was used in design.lisp.  As a consequence, the compiler incorrectly
assumed that WITH-MEDIUM-OPTIONS was an undefined function.  By moving
WITH-MEDIUM-OPTIONS to design.lisp, this problem is solved.

Furthermore, the expansion of WITH-MEDIUM-OPTIONS contains a call to
the internal generic function DO-GRAPHICS-WITH-OPTIONS.  All
definitions of methods on DO-GRAPHICS-WITH-OPTIONS occur in
graphics.lisp, and there is no corresponding DEFGENERIC anywhere.  For
a cleaner compilation, I added a DEFGENERIC form of
DO-GRAPHICS-WITH-OPTIONS in design.lisp.