Included an explicit DEFGENERIC form for TRANSFORMATION-TRANSFORMATOR [sic].
There are two methods on this generic function, but as far as I can
see, the function is nowhere used.  For that reason, I also included a
FIXME comment, saying that Gilbert Baumann should be contacted and
asked whether we should keep, rename, or remove this function and its
methods.