Added explicit DEFGENERIC form for REGION-EXCLUSIVE-OR.
The generic function REGION-EXCLUSIVE-OR is not part of the CLIM II
spec, so there is no explicit DEFGENERIC form in the "Protocol"
section of the file.  However, certain Common Lisp compilers emit a
style warning unless there is such an explicit form.  For that reason,
in order to get a clean build, we include the DEFGENERIC form right
before the (only) DEFMETHOD form for REGION-EXCLUSIVE-OR.