Moved DEFGENERIC of SCROLL-QUANTUM from panes to stream-output.
The file stream-output.lisp belongs to the system named :CLIM-BASIC
and the file panes.lisp belong to the system named :CLIM-CORE.  The
system :CLIM-BASIC is compiled before :CLIM-CORE.  The DEFGENERIC for
SCROLL-QUANTUM used to be in panes.lisp, but there is a DEFMETHOD form
for it in stream-output.lisp.  This situation created a style warning.

I avoided the style warning by moving the DEFGENERIC form to
stream-output.lisp so that the generic function gets is created before
the first method definition for it.