Don't depend on swank, minor CLX fix
The mcclim.asd file can now be loaded in a mode, which makes it not
depend on swank in any form. To use this, push ":clim-without-swank"
onto "*features*".

The CLX backend now passes the ":size 16" parameter do xlib:draw-glyphs
in medium-draw-text*. This was done since forever already be
medium-draw-glyph, but it seems to have been missing from draw-text*.