New directory for improved compilation of the specification.
I made a copy of all the .tex, .epsi, and .lisp files in the Spec
directory, but I didn't copy the .pdf files because they are going to
be built from the .epsi files.

I am hoping this directory will ultimately make the Spec directory
obsolete, but for now I am leaving both of them in there.