Do not skip primary method even when sizes have not changed.
This change is a step in the direction of fixing the fact that
changing the frame layout did not work.  The problem is very likely
not entirely fixed though.  For one thing, LAYOUT-FRAME is not called
when a frame is adopted.  Instead, some more-or-less equivalent
processing is done.  This processing should be moved to LAYOUT-FRAME
and LAYOUT-FRAME should be called instead.