text-style-to-X-font: make function generic
xrender-fonts from the truetype extensions specializes on the port and
text style while the text-style-to-X-font is an ordinary function by
default.

This was hacked out by calling FMAKUNBOUND and defining the
method (which implicitly created a generic function). Hack is removed in
favour of making a function generic and providing a default method.