Don't mandate quoting the type in define-command
Up to now McCLIM didn't accept argument types which weren't quoted:

(define-app-command xyz ((arg symbol)) …)

would signal a condition, that SYMBOL is unbound. On the other hand
other CLIM implementations allowed that. Right now we allow quoting
type, but if it's not quoted, we're not evaluating it anyway.

This is both (almost-)backward-compatible and should make other CLIM
applications compile on McCLIM.