Address issue 35: have mcclim handle keymap change
Basically, I adjusted the code in port.lisp from the CLX backend to
listen for the :mapping-notify X event and, when received, call
(xlib:mapping-notify) with the appropriate arguments.

One note about this commit remains: I'm not sure what the appropriate
argument for the start parameter of mapping-notify is and mapping-notify
ignores that parameter anyways, but declares its type to be a fixnum, so
I just pass 0 instead.