multiprocessing: be on a safe side wrt condition-wait with timeout
If tiemout is NIL, call condition-wait without timeout key
parameter. Should fix at least some problems on ECL and
implementations not supporting timed condition-wait.

If timeout is non-NIL, it must be for a reason, so either fix the
reason in the code calling condition-wait, or error if implementation
doesn't support this.

Fixes #186.