generic-list-pane: handle visible-items more intuitevely
If visible-items isn't set during pane creation, we always
take *actual* number of elements. Up to now we have assigned *initial*
number of elements to it, what resulted in confusing behavior when new
list was assigned as items, which was longer then the initial one -
list length was cut.

Fixes #198.