don't die on asynchronous errors
CLX isn't very thread-safe, but most of the errors are not
critical (i.e shifted response) and doesn't influence overall
effect. When we handle these errors, user is dropped into the
debugger, and closing it breaks X connection.

Workaround is to ignore errors for asynchronous connections. Also we
log each error on with log4cl.