Fix value-changed-callback handling in text-editor-pane
The way this was implemented previously, namely passing the value of
the value-changed-callback /initarg/ to the substrate object, bypassed
the value-changed-callback /method/ of the text-editor-pane instance,
violating the description of value-changed-callback in 30.3 (Basic
Gadget Classes).

With this change a value-changed callback is installed in the
substrate object that invokes the value-changed-callback method for
the text-editor-pane.

Another (potential) benefit of this change is that the id and client
initargs can be handled by the text-editor-pane instance normally
instead of being passed to the substrate object. This seems more
accurate and more robust (e.g. in case the id and/or client were to
change in the text-editor-pane instance for some reason, the new
values would be "picked up").