local-record-p: remove unused special variable
After manual flow analysis I've came to conclusion, that this flag is
not used (maybe it was in the past, but now it's not the case).

stream-write-output is called by:
stream-write-output <- replay-output-record <- replay
stream-write-output <- seos-write-string [ <- seos-write-string ]

stream-write-output calls:
stream-write-output -> stream-add-*-output -> add-*-to-text-record

there is no path leading to recursive call of stream-write-output,
hence I conclude that this flag may be safely removed. Simple tests
indicate no change in application behavior.