recording: extract with-output-recording-options up
we don't use letf for setting stream-recording-p, which is a pattern
duplication for with-output-recording-options.

Moreover we group both operations in handle-repaint under
with-output-recording-options, because they are meant to be tied
together (if we provide locking, this will tie them up), and
stream-replay calls replay which does the same thing anyway.