draw-design: elliptical-thing, polygon and ellipse fixes
elliptical-thing: draw-ellipse* was called with bogus
multiple-value-call. Probably a leftover from a refactor, because
multiple-value-bind is put above it and arguments are layed exaclty as apply
wants.

Due to multiple-value-call last argument options was passed as a list,
so we had odd number of key arguments error when for instance :ink
+blue+ was passed. Fixes #341.

polygon draw-design tried to supply :closed parameter from
method (polyline-closed design), which was wrong, because polygon is not a
polyline (and method is undefined). It is possible to supply closed as one of
rest options to draw-design.

ellipse: provide default values for angles if null