CLX backend: clipping region: provide optimization for ellipses
We keep generic slow code for arbitrary regions, but we provide faster
implementation for ellipses. Generic code is O(n^2) because it iterates over all
pixels in the area, while ellipse code is O(n).

In this optimization we scan vertical lines and got after intersection, since we
know that ellipse is a convex object (with optional limited angle, what may give
us region composed of two lines).