bordered-output: improvements
Closes #165.

- we record stream, sharpe and drawing-options in a slot
- %prepare-bordered-output-record function is simplified
- specialization for replay-output-record which recreates border
- initialize correctly inner-record initial position
- clarify comments in a few places regarding borders and record position

We want to be able to recompute border under and over the record when necessary,
hence we need to record some options. %pbor simplification is a natural
following step.

When we replay output for border, we clean and recreate its children. Thanks to
that we can move children records as in setf output-record-position, or even
resize them, and border will always be drawn properly.

Sometimes we deal with empty body in surrounding. In that case we need to
initialize position of the empty compound-output-record to be sure, we don't
always land in upper left corner.