Assert empty options in with-debugger macro
API has changed recently and this makes it more certain that user will know that
first form is not part of the body now (in already existing code). Fixes #434.

We also bind bt:*default-special-bindings* to ensure, that a subprocess spawned
in with-debugger body will have the debugger as well. (Note that process spawned
from subprocess won't have these default-special-bindings bound).

Clean whitespace.