Get rid of overblown mirror class hierarchy
- collapse all mirror classes into mirrored-sheet-mixin

Backend-specific behavior should be implemented with port implementation, mirror
handling should be uniform.

- remove custom "note-*-mirrored-child" protocol

Notification protocol is for the user (while setf functions are for the
implementation). Merge from 2y ago did not obey this rule.

- move code from "common/mirrored-sheets.lisp" to corresponding modules

This code arrangement didn't make much sense to start with. Next is
standard-event-* and distribute-event.