always-repaint-background-mixin: mix in CLX-fb mirrors
This class is mixed in output-recording-streams and in CLX-fb mirrors (otherwise
we have glitches). There is no reason to mix it *everywhere*.

Some other areas where we could improve:

- instead of *inhibit-dispatch-repaint* we could accumulate regions to repaint
and postpone it until later. That way highlight/unhighlight output-record could
be considerably faster (only one region to repaint in case of single-box=NIL;

- raise called on non-mirrored sheet which is a mirror sibling won't work, not
sure how to achieve that without gross hacks;

- there is a problem with ttf fonts drawing not being idempotent. We should either
fix that or always repaint background over text (but repainting background over
everything is an overkill);

- some (all?) gadgets repaint their background. We could define method
pane-background on them (so the effective-gadget-background is returned) and
simply mix always-repaint-background-mixin.