ux: text-selection: small function refactor and more general approach
We reduce handle-repaint function nesting. Also we get rid of two dynamic
variables (we may reintroduce them back as configuration options in the future
maybe?) in favour of default colors.

Text highlight is achieved twofold: if the colors are default (white/black) we
use light-yellow/dark-red. Otherwise we swap colors. This assures correctness
for odd color settings and looks nice on default ones.