Fix issue with adding a new page to tab-layout.
The method (setf tab-layout-pages) tried to set the sheet-enabled-p field before setting the parent (via sheet-adopt-child). Setting the sheet-enabled-p field called dispatch-paint on the parent as a side effect, but the parent was NIL which failed. Fix is to reverse the order, set the parent first, then set the sheet-enabled-p field.