Remove the inconsistency between :FIX and :FIXED
According to the spec, the keyword :FIX should be used to indicate a
fixed-width font. In various places in the code, the additional
keyword :FIXED was also allowed, and spacial cases were implemented to
transform that to :FIX.

However, this was not consistent, and specifically the implementation
of TEXT-STYLE-FIXED-WIDTH-P returned true only for the case where the
font was set to :FIXED.

This fix removes any reference to :FIXED, in order to make the API
consistent. One of the examples in clim-examples was also changed to
use :FIX instead of :FIXED.