Freetype: Don't crash when coordinate swizzling is active
Sometimes there may be requests to draw text just outside the text
area. If coordinate swizzling is active, this area will be just above
position #xFFFF. Such strings should not be visible, but if an attempt
is made to actually copy these characters to the window, xlib will
throw an exception because the coordinate is out of range.

This fix ensures that no characters are drawn outside the valid range.