Fix repaint order in incremental-redisplay
Incremental-redisplay works by calling compute-difference-set which
collects a list of drawing operations which are to be performed. It
does this by pushing the output records to a list. This results in the
order of the output records being reversed, which causes them to be
drawn in the wrong order. This can behaviour can be seen when adding
two graphics operations to an updating-output form which draws on top
of each other. After redisplay, the output records will have been
drawn in reverse order.

This fix simply calls REVERSE on the list before returning it.