Fix the :CONTENTS initialiser for SHEET-MULTIPLE-CHILD-MIXIN
Previously, the CHILDREN slot has an initarg called :CHILDREN. This
initarg was not usable since none of the logic implemented by
SHEET-ADOPT-CHILD was performed.

This fix changes this so that SHEET-ADOPT-CHILD is called once for
each child. It also renames the keyword argument to :CONTENTS in order
to be consistent with other panes.