invoke-with-room-for-graphics: add naive implementation for medium
For panes not being extended-output-stream it sets beginning at translated [0,0]
point and sets medium transformation to that translation + identity or
Y-inversion depending on first-quadrant key argument.

MOVE-CURSOR and RECORD-TYPE are ignored. Later we'll want to support clipping to
height and returning somehow imaginary-cursor progress on MOVE-CURSOR = T. This
implementation is good enough to make drawing on basic-pane inside
with-room-for-graphics work, though medium chokes on inverted Y-axis and draws
nothing.