Remove old code that was part of the optimised scroll (#519)
* Remove old code that was part of the optimised scroll

The optimised scroll implementation implemented scrolling by copying
the content of the window instead of repaining it. This worked
sometimes, but was problematic when a window was partially visible.

As it turns out, the proper way to do this is to always draw the
window content to a backing pixmap so that the previous content is
always available. If this is done, not much of the old code will be
useful anyway, so to clean things up this commit removes all of the
unused code.

* Remove a comment that doesn't apply anymore