Do not translate window coordinates relative to its parent
When receiving a window configuration event, the code previously
translated the coordinate reelative to its parent. The purpose of this
was to deal with a problem where windows moved slightly after they
were resized.

The idea behind the previous behaviour was that when a window is
created, the window manager moves it into an enclosing window that
manages the window decorations. If the window manager decorations is
10 pixels wide and high, the inner window will be moved 10 pixels to
the down and right. The purpose of the previous code was to avoid this
by taking window decorations into account when placing windows.

As it turned out, there are some issues with this approach. The main
one is that the reason this worked was that it is assumed the window
manager always moves the window after placing it inside the
decoration. However, not all window managers does this, for example
StumpWM when using a floating group.

Another case where this is a problem is when running Qubes OS. In
Qubes, the window manager isn't actually running in the same VM as the
client, and from the point of view of the client, the parent window is
the root window. This causes the window position to be wildly off.

Another issue is that there is no guarantee that window decorations
are contained in the parent window. There may be several windows
involved, and the spec doesn't say anything about how it's supposed to
work.

I, with assistance from Jackdaniel who was the one who found the
original window offset problem to begin with, have attempted to
reproduce the old behaviour that the earlier fix attempted to address,
but we have been unable to do so. Therefore, this fix is introduced
that restores the previous behaviour, thereby fixing all known
problems coming from uncommon window manager behaviour (in particular,
it makes menus draw correctly on Qubes OS, instead of in the
completely wrong location, sometimes so far off that it craches the
application).

It is possible that the original issue of misaligned windows when
resizing that Jackdaniel found still exists, and if it does, I will
need to address that in a more stable way.