make-space-requirement: clamp optimal width/height to min/max bounds
We keep the assertion (<= min max), but min/optimal/max values are clamped
before that (min and max to 0;+fill+ and optimal to min;max).

That will make combining requirements more intuitive (previously, when we
specified only min-width for the component width defaulted to 0, which is
smaller, and dropped us in the debugger). Similarily, max-width/height could end
up being bigger than +fill+ when we called space-requirement+.

We also add min-width and min-height for text-pane to fit at least one
character. New modify macro is introduced in climi called CLAMPF (clamps and
setf's the place).