indirect-ink: make things work smoother
1. with-output-recording-options is sure to invoke things inside
with-drawing-options macro. This is important if we have some pre-set setup for
medium covered in graphics options (i.e setting indirect-ink dynamic variable).

2. do-graphics-with-options binds *foreground-ink* and *background-ink* (this
was the actual bug).

3. Add a comment showing how we could enforce medium's foreground/background if
we really wanted to not use the new protocol. Current solution has a benefit
that we can hack clime:indirect-ink-ink method to return constant (arbitrary)
values for testing.