graphics: INVOKE-WITH-DRAWING-OPTIONS passes correct medium to continuation
This change only affects the case in which MEDIUM is a sheet.

The specification entry for INVOKE-WITH-DRAWING-OPTIONS could imply
that CONTINUATION is called with the same medium object passed to
INVOKE-WITH-DRAWING-OPTIONS. McClim assumes this is the case since it
yields more consistent behavior in case the MEDIUM argument is a sheet
with output recording: in this case, passing e.g. the sheet's medium
with the appropriate drawing options to CONTINUATION instead of
the sheet itself would bypass output recording.

Also changed the parameter name from sheet to medium for the default
method since that is the name used in the specification and in the
generic function.