protocol-classes: remove runtime overhead, restore initarg checking
The previous implementation incurred quite the runtime overhead when
making instances of subclasses of protocol classes: one
INITIALIZE-INSTANCE :AFTER method doing (eq (class-of …) …) for each
protocol class in the superclass list.

The new implementation injects a dummy slot the initfunction of which
signals the "FOO is a protocol class …" error into the protocol class
but not its subclasses. This technique should not incur any runtime
overhead.

This change also solves the problem that the INITIALIZE-INSTANCE
:AFTER method defined for a protocol class specified &ALLOW-OTHER-KEYS
in its lambda list, effectively disabling initarg checking in its
subclasses.

Benchmark:

  (defun foo ()
    (loop :repeat 10000000 :do (make-instance 'standard-bounding-rectangle
                                              :x1 1 :y1 2 :x2 3 :y2 4)))
  (time (foo))

  ;;; Before
  Evaluation took:
    4.406 seconds of real time
    4.381687 seconds of total run time (4.313407 user, 0.068280 system)
    [ Run times consist of 0.640 seconds GC time, and 3.742 seconds non-GC time. ]
    99.46% CPU
    13,187,450,016 processor cycles
    2,728,196,160 bytes consed

  ;;; After
  Evaluation took:
    1.982 seconds of real time
    1.955515 seconds of total run time (1.918418 user, 0.037097 system)
    [ Run times consist of 0.633 seconds GC time, and 1.323 seconds non-GC time. ]
    98.69% CPU
    5,931,761,385 processor cycles
    2,728,154,576 bytes consed