cleanup: remove stream-wrap-line protocol
This was an unnecessary hack - we were storing margin value in recorded text for
sake of output-record-equal on text-output-record. Since we compare text
bounding-box (via text-bounding-rectangle*), then wrapped is unnecessary and
this removal slightly simplifies the architecture.

nb: wrapped was also used for a hack with drawing a rectangle on line wrap what
worked only for first draw, it wasn't replayed. I assume it was a test code
which haven't got disabled.