stream-output: make seos work with streams w/o bounding rectangle
+everywhere+ region doesn't have bounding-rectangle-width nor
bounding-rectangle-height, so there is no default margin value nor page height.

When margin=NIL and region=+everywhere+ we return +fill+. I'd rather return NIL
and update the spec for that, but such attempt will break the contract with a
programmer (our own code assumes in many places that it is a number and we can
expect that other code in the wild makes a similar assumption).

We also remove a disabled hack which I have no idea what it was there for.