clime: unify pane/stream/medium-background/foreground
We leave CLIM-official readers (pane- and medium-), but all also have a reader
clime:background and clime:foreground for simplicity. We also fix the call in
text-drawing-utils.

line-style-effective-thickness is not exported from clime package. When we
implement a protocol in clim-backend package which spans all backends we'll
think about re-exporting it from clim-extensions.