backend: common: add climb:approx-bbox-medium-mixin
This medium is responsible for fixing the bounding box for any alignment. It is
approximation method, because it takes whole bbox and manipulates it. In comment
we have explanation of a fail-case.

Also we change interface so climb:text-bounding-rectangle* and
climb:font-text-extents accepts align-x/align-y to be able to provide exact
methods.