Registe xkb.lisp as a component of mcclim-clx
For the time being, let XKB be a part of mcclim-clx itself.  Actually,
the support for XKB could be a subsystem of mcclim-clx, but then the
defaults in port.lisp override those in xkb.lisp.  Alternatively, it
could be a supersystem of mcclim-clx, since XKB is an *extension*, but
then XKB wouldn't work out of the box.