gadgets: rewrite BOX-ADJUSTER-GADGET
This re-implementation uses a different approach: instead of tweaking
the space requirements of the left and right client, the box client
strategies (fixed size, proportional size, fill) are adjusted in a way
that preserves the general strategy (e.g. a client using fill above a
client using proportional) but changes the parameters.

Furthermore, instead of tracking an offset from the start point of the
drag, the pointer position in the coordinate system of the left child
is tracked. This is necessary for the implementation strategy
described above.

Other changes:

* Use :move cursor instead of :rotate cursor

* Introduce constants for magic numbers, add comments

* Store transient dragging state in a separate object instead of
  several slots