panes: be more roboust with default values
top-level-sheet-pane's native transformation could be NIL (what was incorrect
given NIL is not a transformation). That lead to errors in
update-mirror-geometry.

basic-pane had no default width/height, so if we had put simply

  (make-pane 'application-pane)

as pane option we had 0 width/height what lead to errors in clx (bad requests
for windows of zero size). Fixes problem reported in wrong repository:

https://github.com/sharplispers/clx/issues/123