input: event-queue: refactor module
- extract schedule experimental timer into a mixin

Also change its "generic functions" to ordinary functions. There was absolutely
no need for them.

- remove subclass port-event-queue

We have *never* used standard-event-queue, only its subclass
port-event-queue. We just pull the introduced slot higher and have it there.

- provide internal protocol class event-queue

I was tempted to replace all event-queue-* generic functions with defuns because
this is a singleton protocol, but I've decided to keep it and have a protocol
class for it. Protocol is cleaned up a little (it could be cleaned more!)

- export in clim-extensions event-read-with-timeout and schedule-event

Both are clearly meant as extensions for missing functionalities in the spec.