event-queue: protect setting port of a shared queue to NIL
Current practice in McCLIM is to share the same queue by default between frames
in parent-children relation. That means we can't setf event-queue to NIL on
frame-disown - otherwise parent will have problems with accessing events if it
calls process-next-event. We provide basic protection for a case when parent and
child queues are eql (we don't erase port in that case).

We also remove a singleton protocol pane-enabledp - it is not used anywhere for
any purpose in the codebase nor it is mentioned in the spec. We have sheet
protocol for judging whenever it is enabled. It is a single line removal without
any side effects, so it doesn't deserve its own commit.