Make PARSE-SPACE handle all real numbers (#645)
Previously PARSE-SPACE will raise an error if the specification was a
non-integer number. This was the case even if the number is a floating
point number with a zero fractional part. The error is raised by a
call to ETYPECASE which checks for an integer.

This can happen if one sets a font size to fractional size and
attempts to display a FORMATTING-TABLE. The implementation of
FORMATTING-TABLE reads the STREAM-VERTICAL-SPACING, which is now a
floating-point number and passes it to PARSE-SPACE.

This fix changes the check from INTEGER to REAL so that any kind of
number is supported.