simple-event-queue: do preemptive checks where appropriate
We did check the queue for events and if not present we test for the timeout. If
timeout is not reached we process-next-event. This was wrong in a scenario,
where we have some events available on the backend but still not processed. We
do a similar preemptive check for event-queue-read-no-hang. Added checks have
timeout 0 so they don't block.