In DESIGN-INK, use consistent rounding for coordinates
Previously, ROUND was used to ensure that the arguments to DESIGN-INK
were integers. This caused a problem in some situations when the
coordinates have a fractional part of 0.5. In such cases, ROUND will
either round up or down depending on whether the integer part is odd
or even.

Fixes https://github.com/McCLIM/McCLIM/issues/641