standard-output-recording-stream: don't draw text eagerly
To maintain a baseline (see 15.3 The Text Cursor) we need to see into the future
about other baselines on the line. While this is not possible for streams which
are not recorded it can be done for recording streams.

It is important to remember that we can't immedietely draw text but rather wait
for line completion (or to be more exact, for output record completion which is
finialized with STREAM-CLOSE-TEXT-OUTPUT-RECORD. Otherwise during initial
display result is different on a screen than after repaint.

Additionally we avoid reaching back to STREAM-WRITE-OUTPUT from
REPLAY-OUTPUT-RECORD to skip unnecessary roundtrip (and possible recursion).

CLIM-STREAM-PANE inheritance order is changed: STANDARD-OUTPUT-RECORDING-STREAM
methods take precedence over STANDARD-EXTENDED-OUTPUT-STREAM
ones. OUTPUT-RECORDING-STREAMs abstraction-wise are above
EXTENDED-OUTPUT-STREAMs. This also allows us to drop :around method in favor of
primary method on STREAM-WRITE-OUTPUT.

Fixes #667.