dialog-views: define gadget views for dialogs
With this commit the views for using gadgets in
accepting-values (subclasses of GADGET-VIEW) have the same
initargs of the corresponding gadget  (as written in
Franz User Manual (CLIM 2.2) sec. 8.6.13).

These views have also two class-allocated slots: one with the
gadget-name symbol and the other with the name and initarg of the
gadget slots.

It's also defined the default presentation method for
ACCEPT-PRESENT-DEFAULT and ACCEPT for these views that fallback to the
methods defined for TEXTUAL-DIALOG-VIEW.