seos-write-string: implement better line wrap algorithms
Normal line wrap (:END-OF-LINE-ACTION :WRAP) is implemented with bisection
instead of traversing collection from start to end.

New option (:END-OF-LINE-ACTION :WRAP*) which implements greedy word wrap (also
with bisection but only on spaces). In principle adding unicode wrap should be
easy with adjusting predicate in BREAK-OPPORTUNITY-INDEXES local function -
computing these indexes is not so obvious though.

We do not leave "chsize" for the margin. I think it is a glitch from older
implementation where on line wrap a black rectangle was drawn.