dialog-views: views store initargs instead of mirroring slots
The previous solution defined view classes such that the slots
mirrored the slots of the corresponding view classes. Since all
effective slots must be mirrored (there is no way to know which slots
to include) the approach leads to problems with slots that the gadget
needs but which don't make sense in the view class as well as slot
type checking and initialization.

As a fix, store initargs supplied to view instances as a list in the
instance, making sure view classes only accept initargs also accepted
by the slot class (this is still a superset of the initargs that are
actually useful for views). This avoids problems with slot type
checking and initialization while still somewhat restricting the set
of accepting initargs.