clim-stream-pane: provide default width/height = 100/100
X protocol can't handle windows of width/height = 0. It can handle them when
size is 1x1, but this doesn't make much sense as a default. We provide defaults
for 100x100. Note, that stream-width/stream-height are somewhat inconsistent
with default width/height intiargs despite some visible effort to make them as
an extension of said mechanism. Fixes #47.