Scigraph: remove accept from dwim package
The ACCEPT version in DWIM package produced a SHEET-IS-NOT-A-CHILD
error when a dialog with gadget was closed.

This is due to how QUERY-IDENTIFIER parameter was passed to the
CLIM:ACCEPT function.

Because the Scigraph version in McCLIM is a CLIM2 only application the
DWIM:ACCEPT compatibility layer is no more necessary.

For this reason this commit remove the function ACCEPT from DWIM
package and now Scigraph use directly the CLIM:ACCEPT function.