(setf frame-current-layout): do not suggest width and height
Suggesting width/height based on current frame geometry misses the
point, because current geometry was computed fro the previous
layout. Case in point:

We have narrow (default) and wide layout, then we switch to a wide
one (size is recomputed because it must to fit the stream), and then
we switch back. In this case size of wide layout is preserved while it
should shrink back to the original size. This commit makes that
happen.