%line-break-1: handle gracefully splits with 0 width
Issue is reproducible if the page region is +nowhere+ (page regions
are introduced in further commits). We correctly identify the
initial-break but it was not guaranteed it is part of break-line
collection. This fix is not necessary for %line-break-2 because the
latter calls %line-break-1 if it can't fit the word in an empty line.