Clipboard redesign: Implementation of CLX backend
This patch implements CLX support for the new clipboard framework.

The main user-visible change is that after selecting text by
shift-dragging, it is now possible to paste that text into another
application, using a middle click to paste the selection. Pasting
from the clipboard will come in a later patch.