Clipboard redesign: Implement copy commands
This patch introduces two new commands in the global command table:

"Copy Presentation to Clipboard" accepts a presentation which contains
a copyable object (i.e. an object for which CONVERT-CLIPBOARD-CONTENT
returns non-nil), and places that object on the clipboard.

"Copy Text to Clipboard" is the CLIM version of the control-c in most
applications. It copies the currently highlighted text (i.e. the local
selection) and copies it as a string to the clipboard.