text-formatting: filling-output: allow composing line-break prints
Additionally we remember the indent from where filling-output was
called. That means in particular that it is always lined up to the
current initial position. In a worst case after-line-break may
overlap. Introduces a new keyword :after-line-break-composed
defaulting to T.

We get rid of a state slot in favor of a closure for ink and
text-style.