page-layout: margins: recompute margins eagerly
When we set margins we recompute them right away. This change is
necessary for filling-output invoked over indenting-output work
correctly with regard wrapping (nested with-temporary-margins). As of
why this change is necessary I'm not sure (from the program flow both
variants should be equivalent), but I don't have energy to investigate
that provided that improved version just works™.

Also wrap move-cursor in invoke-with-temporary-page in unwind-protect.