standard-page-layout: margins: use initform instead of initarg
Slot is specified as of a type margin-spec. We fix the incomplete
specifications in initialize-instance :after method. If the slot had
initarg, then said incomplete specification is first set to the slot
what violates its type specification (what leads to error when high
enough debug settings are set). Moreover (setf stream-text-margin)
accesses the old margin, so we can't leave the slot unbound.