stream-cursor-initial-position: fix a few regressions
Since margin introduction we can't assume that page starts at cursor
position (values 0 0). Drei seemed to make this assumption what kind
of broke interactor - spurious new lines from read-command. We fix the
cursor position in pdf and ps backends too since we're at it.

Moreover we adjust margins for with-output-to-output-record to match
the viewport (spotting this issue was a result of investigating #737.

Closes #737.