cleanup: remove cut-and-paste-mixin and selection api
We remove asynchronous API to add a new one which will be synchronous
and will support different selections (more general protocol).

cut-and-paste-mixin is removed now but it will be re-added as
text-selection-mixin.

Interaction Shift-Mouse-M for paste on seos is removed because it
doesn't make sense in a first place, it pastes selected text on
arbitrary panes. This interaction will be added back for text-fields
in Drei with a new selection protocol