repaint-sheet: don't call bounding-rectangle on passed region
We do a very careful work to cut everything outside the mirror so
there are no out-of-coordinate operations (16 bit on X11) but then we
simply called bounding-rectangle on resulting region. In case of
nowhere-region that yield bounding rectangle [0:0] [0:0]. If this
point is not part of the sheet native region (due to coordinate
swizzling) we have an error. Fixes #743.