This makes the menu behave like other menu's.
An open menu will now no longer close during the first click when the mouse-button is released unless it is released over a button with a command.
Clicking on an open menu will close it, as will clicking outside of the menu.

To support this, it was necessary to track which menu was open.
The active menu is now tracked on the standard-application-frame in a slot named 'active-menu'.