pointer-position and stream-pointer-position
With this commit POINTER-POSITION return the x y position of the
pointer in the graft coordinate system.

This is an interpretation of the clim specifications (section 22.4)
different from what is written in Franz Manual (section 15.1) and
different from what McCLIM have done until now.

For Franz manual and in McCLIM before this commit, POINTER-POSITION
return the position of the pointer in coordinate system of the sheet
that the pointer is currently over.

I find the choice of this commit better for the following reasons:

1. more logical (personal opinion)

2. In the Franz choice it is not clear what values are returned if the
   pointer isn't over a clim sheet.

3. It is not necessary to investigate about what is the sheet that the
   pointer is currently over to interpret the returned values.

4. The sheet that the pointer is over can be very volatile (like menu
   panes) and can not exist when the values returned by the function
   is used.

Other changes in this commit are:

The STREAM-POINTER-POSITION is redefined using the new
POINTER-POSITION function.

The SHEET-POINTER-POSITION generic function is removed because it
isn't used anymore.

Because the semantic of POINTER_POSITION is changed I checked when the
function was called. I found that in McCLIM code POINTER-POSITION was
called only once in DRAGGING-DRAWING function but the returned values
was never used so I have removed the call.